+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## next
+
+- Add readyStates as both static and and instance members
+
 ## 1.0.0
 
 Initial release: